backend/requirements.txt
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
Pillow
